feat: Remove custom bound checkboxes
