+++ Attributes/ValidationAttributes/GovUkValidateCheckboxNumberOfResponsesRangeAttribute.cs
@@ -1,19 +1,21 @@
 ﻿using System;
+using System.Collections;
+using System.ComponentModel.DataAnnotations;
 
 namespace GovUkDesignSystem.Attributes.ValidationAttributes
 {
     [AttributeUsage(AttributeTargets.Property)]
-    public class GovUkValidateCheckboxNumberOfResponsesRangeAttribute : Attribute
+    public class GovUkValidateCheckboxNumberOfResponsesRangeAttribute : ValidationAttribute
     {
         /// <summary>
         ///     The minimum number of checkboxes the user must select for us not to show an error
-        ///     <br/>If this is omitted, or has the value null, then there is no minimum
+        ///     <br/>If this is omitted then there is no minimum
         /// </summary>
         public int MinimumSelected { get; set; } = int.MinValue;
 
         /// <summary>
         ///     The maximum number of checkboxes the user can select before we show an error
-        ///     <br/>If this is omitted, or has the value null, then there is no maximum
+        ///     <br/>If this is omitted, then there is no maximum
         /// </summary>
         public int MaximumSelected { get; set; } = int.MaxValue;
 
@@ -34,5 +36,32 @@ namespace GovUkDesignSystem.Attributes.ValidationAttributes
         /// </summary>
         public string ErrorMessageIfNothingSelected { get; set; }
 
+        /// <summary>
+        /// The name to use in error messages about the number of selected options
+        /// </summary>
+        public string PropertyNameForErrorMessage { get; set; }
+
+        protected override ValidationResult IsValid(object value, ValidationContext validationContext)
+        {
+            var selectedValues = (IList)value;
+
+            if (selectedValues.Count == 0 &&
+                !string.IsNullOrEmpty(ErrorMessageIfNothingSelected))
+            {
+                return new ValidationResult(ErrorMessageIfNothingSelected);
+            }
+
+            if (selectedValues.Count < MinimumSelected)
+            {
+                return new ValidationResult($"Select at least {MinimumSelected} options for {PropertyNameForErrorMessage}");
+            }
+
+            if (selectedValues.Count > MaximumSelected)
+            {
+                return new ValidationResult($"Select at most {MaximumSelected} options for {PropertyNameForErrorMessage}");
+            }
+
+            return ValidationResult.Success;
+        }
     }
 }
